vfs: Replace uaccess_[enable|disable]_privileged() with __uaccess_[disable|enable]_hw_pan()
Older kernels like 5.4 does not have uaccess_[enable|disable]_privileged() out of the box.
And the function set has various sanity checks which prove to be unneeded based on our
configuration of the kernel. If we import only the necessary parts of the void function
based on our configuration which has CONFIG_ARCH_MTE and CONFIG_ARM64_SW_TTBR0_PAN, we
end up with __uaccess_[disable|enable_hw_pan(). Therefore, switch to them in order to
fix compilation errors on older kernel revisions.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>